move code from `utils.rs` to `schema_inference.rs`
